API

Public

SimJulia — Module.

Main module for SimJulia.jl – a combined continuous time / discrete event process oriented simulation framework for Julia.

source

Events

SimJulia.AbstractEvent — Type.

The parent type for all events.

An events holds a pointer to an instance of a subtype of Environment.

An event has a state:

may happen (idle),

is going to happen (scheduled),

has happened (triggered).

Once the events is scheduled, it has a value.

An event has also a list of callbacks. A callback can be any function as long as it accepts an instance of a subtype of AbstractEvent as its first argument. Once an event gets triggered, all callbacks will be invoked. Callbacks can do further processing with the value it has produced.

source

SimJulia.Timeout — Type.

An event that gets triggered after a delay has passed.

This event is automatically scheduled when it is created.

Signature:

Timeout{E<:Environment} <: AbstractEvent{E}

Field:

bev :: BaseEvent{E}

Constructors:

Timeout{E<:Environment}(env::E, delay::Period; priority::Bool=false, value::Any=nothing)

Timeout{E<:Environment}(env::E, delay::Number=0; priority::Bool=false, value::Any=nothing)

source

Simulation

SimJulia.Simulation — Type.

Execution environment for a simulation.

The passing of time is implemented by stepping from event to event.

Signature: Simulation{T<:TimeType} <: Environment

Fields:

time :: T

heap :: PriorityQueue{BaseEvent{Simulation{T}}, EventKey{T}}

eid :: UInt

sid :: UInt

active_proc :: Nullable{Process}

Constructors:

Simulation{T<:TimeType}(initial_time::T) :: Simulation{T}

Simulation(initial_time::Number=0) :: Simulation{SimulationTime}

An initial_time for the simulation can be specified. By default, it starts at 0.

source

Base.Dates.now — Method.

Returns the current simulation time.

Method: now{T<:TimeType}(sim::Simulation{T}) :: T

source

Base.run — Method.

Executes step until the given criterion is met:

if nothing is not specified, the method will return when there are no further events to be triggered

if it is a subtype of AbstractEvent, the simulation will continue stepping until this event has been triggered and will return its value

if it is a subtype of TimeType, the simulation will continue stepping until the simulation’s time reaches until

if it is a subtype of Period, the simulation will continue stepping during the given period

if it is a subtype of Number, the method will continue stepping during a period of elementary time units

In the first two cases, the simulation can prematurely stop when there are no further events to be triggered.

If the stepping end with a StopSimulation exception the function return the value of the exception, in all other cases the exception is rethrown.

Methods:

run(sim::Simulation, until::AbstractEvent) :: Any

run{T<:TimeType}(sim::Simulation{T}, until::T) :: Any

run(sim::Simulation, period::Union{Period, Number}) :: Any

run(sim::Simulation) :: Any

source

Processes

SimJulia.Process — Type.

A Process is an abstraction for an event yielding function, i.e. a process function.

The process function can suspend its execution by yielding an instance of AbstractEvent. The Environment will take care of resuming the process function with the value of that event once it has happened. The exception of failed events is also thrown into the process function.

A Process is a subtype of AbstractEvent. It is triggered, once the process functions returns or raises an exception. The value of the process is the return value of the process function or the exception, respectively.

Signature:

Process{E<:Environment} <: AbstractEvent{E}

Fields:

bev :: BaseEvent{E}

task :: Task

target :: AbstractEvent{E}

resume :: Function

Constructor:

Process{E<:Environment}(func::Function, env::E, args::Any...)

source

SimJulia.@Process — Macro.

Creates a Process with process function func having a required argument env, i.e. an instance of a subtype of Environment, and a variable number of arguments args....

Signature:

@Process func(env, args...)

source

Base.yield — Method.

Passes the control flow back to the simulation. If the yielded event is triggered, the Environment will resume the function after this statement.

The return value is the value from the yielded event.

Method:

yield(target::AbstractEvent) :: Any

source

Internals

Simulation

SimJulia.step — Method.

Does a simulation step and processes the next event.

Method:

step(sim::Simulation) :: Bool

source