API

Public

SimJulia — Module.

SimJulia

Main module for SimJulia.jl – a combined continuous time / discrete event process oriented simulation framework for Julia.

source

Base

SimJulia.Event — Type.

Event

An event is a state machine with three states:

idle

triggered

processing

Once the processing has ended, the event returns to an idle state and can be scheduled again.

An event is initially not triggered. Events are scheduled for processing by the simulation after they are triggered.

An event has a list of callbacks and a value. A callback can be any function. Once an event gets processed, all callbacks will be invoked. Callbacks can do further processing with the value it has produced.

Failed events, i.e. events having as value an Exception, are never silently ignored and will raise this exception upon being processed.

Fields:

callbacks :: Vector{Function}

state :: EVENT_STATE

value :: Any

Constructor:

Event()

Event(sim::Simulation, delay::Period; priority::Bool=false, value::Any=nothing)

Event(sim::Simulation, delay::Number=0; priority::Bool=false, value::Any=nothing)

source

SimJulia.Simulation — Type.

Simulation{T<:TimeType}

Execution environment for a simulation. The passing of time is implemented by stepping from event to event.

Fields:

time :: T

heap :: PriorityQueue{Event, EventKey}

sid :: UInt

Constructor:

Simulation{T<:TimeType}(initial_time::T) Simulation(initial_time::Number=0)

An initial_time for the simulation can be specified. By default, it starts at 0.

source

Base.Dates.now — Method.

now(sim::Simulation) :: TimeType

Returns the current simulation time.

source

Base.run — Function.

run(sim::Simulation, until::Event)

run(sim::Simulation, until::TimeType)

run(sim::Simulation, until::Period)

run(sim::Simulation, until::Number)

run(sim::Simulation)

Executes step until the given criterion until is met:

if it is not specified, the method will return when there are no further events to be processed

if it is an Event, the method will continue stepping until this event has been triggered and will return its value

if it is a TimeType, the method will continue stepping until the simulation’s time reaches until

if it is a Period, the method will continue stepping until the simulation’s time has passed until periods

if it is a Number, the method will continue stepping until the simulation’s time has passed until elementary periods

In the last two cases, the simulation can prematurely stop when there are no further events to be processed.

source

Base.schedule — Method.

schedule(sim::Simulation, ev::Event, delay::Period; priority::Bool=false, value::Any=nothing) :: Event

schedule(sim::Simulation, ev::Event, delay::Number=0; priority::Bool=false, value::Any=nothing) :: Event

Schedules an event at time sim.time + delay with a priority and a value.

If the event is already scheduled or is being processed, an EventNotIdle exception is thrown.

source

SimJulia.append_callback — Method.

append_callback(ev::Event, cb::Function, args...; include_event::Bool=false, sticky::Bool=false) :: Function

Adds a callback function, i.e. a function having as first argument an object of type Simulation, to the event. The second argument is the event if include_event=true. Optional arguments can be specified by args.... The sticky keyword argument allows to keep a callback function when reusing an event. The default behavior is to remove the callback functions at the end of the processing.

If the event is being processed an EventProcessing exception is thrown.

Callback functions are called in order of adding to the event.

source

SimJulia.schedule! — Method.

schedule!(sim::Simulation, ev::Event, delay::Period; priority::Bool=false, value::Any=nothing) :: Event

schedule!(sim::Simulation, ev::Event, delay::Number=0; priority::Bool=false, value::Any=nothing) :: Event

Schedules an event at time sim.time + delay with a priority and a value.

If the event is already scheduled, the key is updated with the new delay and priority. The new value is also set.

If the event is being processed, an EventProcessing exception is thrown.

source

SimJulia.state — Method.

state(ev::Event) :: EventState

Returns the state of the event.

source

SimJulia.value — Method.

value(ev::Event) :: Any

Returns the value of the event.

source

Processes

Continuous

Resources

Internals

SimJulia.EVENT_STATE — Type.

EVENT_STATE

Enum with values:

idle=0

triggered=1

processing=2

source

SimJulia.EventKey — Type.

EventKey

Key for the event heap.

Fields:

time :: Float64

priority :: Bool

id :: UInt

Constructor:

EventKey(time :: Float64, priority :: Bool, id :: UInt)

Only used internally.

source

SimJulia.EventNotIdle — Type.

EventNotIdle <: Exception

Exception thrown when an event is scheduled (schedule) that has already been scheduled or is being processed.

Only used internally.

source

SimJulia.EventProcessing — Type.

EventProcessing <: Exception

Exception thrown:

when a callback is added to an event (append_callback) that is being processed or

when an event is scheduled (schedule!) that is being processed.

Only used internally.

source

SimJulia.StopSimulation — Type.

StopSimulation <: Exception

Exception that stops the simulation. A return value can be set.

Fields:

value :: Any

Constructor:

StopSimulation(value::Any=nothing)

source

Base.isless — Method.

isless(a::EventKey, b::EventKey) :: Bool

Compairs two EventKey. The criteria in order of importance are:

time of processing

priority when time of processing is equal

scheduling id, i.e. the event that was first scheduled is first processed when time of processing and priority are identical

Only used internally.

source

SimJulia.step — Method.

step(sim::Simulation) :: Bool

Does a simulation step and processes the next event.

Only used internally.

source