API

Public

SimJulia — Module.

SimJulia

Main module for SimJulia.jl – a combined continuous time / discrete event process oriented simulation framework for Julia.

source

Base

SimJulia.EVENT_IDLE — Constant.

const EVENT_IDLE

State representing an event that may happen but is not yet scheduled.

source

SimJulia.EVENT_PROCESSING — Constant.

const EVENT_PROCESSING

State representing an event that is happening.

source

SimJulia.EVENT_TRIGGERED — Constant.

const EVENT_TRIGGERED

State representing an event that is going to happen, i.e. is scheduled but processing has not yet been started.

source

SimJulia.Event — Type.

Event

An event is a state machine with three states:

EVENT_IDLE

EVENT_TRIGGERED

EVENT_PROCESSING

Once the processing has ended, the event returns to an EVENT_IDLE state and can be scheduled again.

An event is initially not triggered. Events are scheduled for processing by the simulation after they are triggered.

An event has a list of callbacks and a value. A callback can be any function. Once an event gets processed, all callbacks will be invoked. Callbacks can do further processing with the value it has produced.

Failed events, i.e. events having as value an Exception, are never silently ignored and will raise this exception upon being processed.

Fields:

callbacks :: Vector{Function}

state :: UInt

value :: Any

Constructor:

Event()

Event(sim::Simulation, delay::Float64; priority::Bool=false, value::Any=nothing)

source

SimJulia.Simulation — Type.

Simulation

Execution environment for a simulation. The passing of time is implemented by stepping from event to event.

Fields:

time :: Float64

heap :: PriorityQueue{Event, EventKey}

sid :: UInt

Constructor:

Simulation(initial_time::Float64=0.0)

An initial_time for the simulation can be specified. By default, it starts at 0.0.

source

SimJulia.StopSimulation — Type.

StopSimulation <: Exception

Exception that stops the simulation. A return value can be set.

Fields:

value :: Any

Constructor:

StopSimulation(value::Any=nothing)

source

Base.Dates.now — Method.

now(sim::Simulation) :: Float64

Returns the current simulation time.

source

Base.run — Function.

run(sim::Simulation, until::Event)

run(sim::Simulation, until::Float64)

run(sim::Simulation)

Executes step until the given criterion until is met:

if it is not specified, the method will return when there are no further events to be processed

if it is an Event, the method will continue stepping until this event has been triggered and will return its value

if it is a Float64, the method will continue stepping until the environment’s time reaches until

In the last two cases, the simulation can prematurely stop when there are no further events to be processed.

source

SimJulia.append_callback — Method.

append_callback(ev::Event, cb::Function, args...)

Adds a callback function to the event. Optional arguments to the callback function can be specified by args.... If the event is being processed an EventProcessing exception is thrown.

Callback functions are called in order of adding to the event.

source

SimJulia.state — Method.

state(ev::Event) :: EventState

Returns the state of the event.

source

SimJulia.value — Method.

value(ev::Event) :: Any

Returns the value of the event.

source

Processes

Continuous

Resources

Internals

SimJulia.EventKey — Type.

EventKey

Key for the event heap.

Fields:

time :: Float64

priority :: Bool

id :: UInt

Constructor:

EventKey(time :: Float64, priority :: Bool, id :: UInt)

Only used internally.

source

SimJulia.EventNotIdle — Type.

EventNotIdle <: Exception

Exception thrown when an event is scheduled (schedule) that has already been scheduled or is being processed.

Only used internally.

source

SimJulia.EventProcessing — Type.

EventProcessing <: Exception

Exception thrown:

when a callback is added to an event (append_callback) that is being processed or

when an event is scheduled (schedule!) that is being processed.

Only used internally.

source

Base.isless — Method.

isless(a::EventKey, b::EventKey) :: Bool

Compairs two EventKey. The criteria in order of importance are:

time of processing

priority when time of processing is equal

scheduling id, i.e. the event that was first scheduled is first processed when time of processing and priority are identical

Only used internally.

source

Base.schedule — Method.

schedule(sim::Simulation, ev::Event, delay::Float64=0.0; priority::Bool=false, value::Any=nothing) :: Event

Schedules an event at time sim.time + delay with a priority and a value.

If the event is already scheduled or is beign processed, an EventNotIdle exception is thrown.

source

SimJulia.schedule! — Method.

schedule!(sim::Simulation, ev::Event, delay::Float64=0.0; priority::Bool=false, value::Any=nothing) :: Event

Schedules an event at time sim.time + delay with a priority and a value.

If the event is already scheduled, the key is updated with the new delay and priority. The new value is also set.

If the event is being processed, an EventProcessing exception is thrown.

source

SimJulia.step — Method.

step(sim::Simulation) :: Bool

Does a simulation step and processes the next event.

Only used internally.

source